+++ compiler-core/src/hex.rs
@@ -51,10 +51,11 @@ pub async fn publish_package<Http: HttpClient>(
     release_tarball: Vec<u8>,
     api_key: &str,
     config: &hexpm::Config,
+    replace: bool,
     http: &Http,
 ) -> Result<()> {
     tracing::info!("Creating API key with Hex");
-    let request = hexpm::publish_package_request(release_tarball, api_key, config);
+    let request = hexpm::publish_package_request(release_tarball, api_key, config, replace);
     let response = http.send(request).await?;
     hexpm::publish_package_response(response).map_err(Error::hex)
 }