+++ compiler-core/src/javascript/tests/bit_strings.rs
@@ -34,6 +34,61 @@ fn go() {
 }
 
 #[test]
+fn integer() {
+    assert_js!(
+        r#"
+fn go() {
+  <<256:int>>
+}
+"#,
+    );
+}
+
+#[test]
+fn float() {
+    assert_js!(
+        r#"
+fn go() {
+  <<1.1:float>>
+}
+"#,
+    );
+}
+
+#[test]
+fn sized() {
+    assert_js!(
+        r#"
+fn go() {
+  <<256:4>>
+}
+"#,
+    );
+}
+
+#[test]
+fn explicit_sized() {
+    assert_js!(
+        r#"
+fn go() {
+  <<256:size(4)>>
+}
+"#,
+    );
+}
+
+#[test]
+fn variable_sized() {
+    assert_js!(
+        r#"
+fn go(x, y) {
+  <<x:size(y)>>
+}
+"#,
+    );
+}
+
+#[test]
 fn variable() {
     assert_js!(
         r#"
@@ -72,6 +127,28 @@ fn bit_string() {
         r#"
 fn go(x) {
   <<x:bit_string, "Gleam":utf8>>
+}
+"#,
+    );
+}
+
+#[test]
+fn empty_match() {
+    assert_js!(
+        r#"
+fn go(x) {
+  assert <<>> = x
+}
+"#,
+    );
+}
+
+#[test]
+fn match_bytes() {
+    assert_js!(
+        r#"
+fn go(x) {
+  assert <<1, y>> = x
 }
 "#,
     );