+++ compiler-wasm/src/log_telemetry.rs
@@ -23,4 +23,20 @@ impl Telemetry for LogTelemetry {
 
         tracing::warn!("Warning: {:?}", string);
     }
+
+    fn downloading_package(&self, name: &str) {
+        tracing::info!("Downloading package: {}", name);
+    }
+
+    fn resolving_package_versions(&self) {
+        tracing::info!("Resolving package versions");
+    }
+
+    fn packages_downloaded(&self, _start: std::time::Instant, count: usize) {
+        tracing::info!("Downloaded {} packages", count);
+    }
+
+    fn waiting_for_build_directory_lock(&self) {
+        tracing::info!("Waiting for build directory lock");
+    }
 }